(subscriptions): use decimal for money values rather than double (#117)

+++ ODK.Services/Chapters/Models/ChapterSubscriptionCreateModel.cs
@@ -2,7 +2,7 @@
 
 public class ChapterSubscriptionCreateModel
 {
-    public required double Amount { get; init; }
+    public required decimal Amount { get; init; }
 
     public required string Description { get; init; }
